chore: migrate user effect components to group effects
Migrate individual user effect components to group effects for better
organization. Retain userSessionResetEffect in its current structure due to
unique requirements.

Update path from tailwind.config.ts to prettier.config.js for consistency.
